remove some shared code
