+++ server/connectors/api-connector-jdbc/src/main/scala/com/prisma/api/connector/jdbc/impl/DeleteNodeInterpreters.scala
@@ -6,7 +6,7 @@ import com.prisma.api.connector.jdbc.database.JdbcActionsBuilder
 import com.prisma.api.schema.APIErrors
 import com.prisma.api.schema.APIErrors.NodesNotConnectedError
 import com.prisma.gc_values.IdGCValue
-import com.prisma.shared.models.{Model, RelationField, Schema}
+import com.prisma.shared.models.{Model, Project, RelationField, Schema}
 import slick.dbio._
 
 import scala.concurrent.ExecutionContext
@@ -24,7 +24,7 @@ case class DeleteNodeInterpreter(mutaction: TopLevelDeleteNode, shouldDeleteRela
                case Some(node) =>
                  for {
                    _ <- performCascadingDelete(mutationBuilder, mutaction.where.model, node.id)
-                   _ <- checkForRequiredRelationsViolations(mutationBuilder, node.id)
+                   _ <- SharedDelete.checkForRequiredRelationsViolations(mutaction.project, mutaction.model, mutationBuilder, Vector(node.id))
                    _ <- mutationBuilder.deleteNodeById(mutaction.where.model, node.id, shouldDeleteRelayIds)
                  } yield node
                case None =>
@@ -32,12 +32,18 @@ case class DeleteNodeInterpreter(mutaction: TopLevelDeleteNode, shouldDeleteRela
              }
     } yield DeleteNodeResult(node, mutaction)
   }
+}
 
-  private def checkForRequiredRelationsViolations(mutationBuilder: JdbcActionsBuilder, id: IdGCValue): DBIO[_] = {
-    val fieldsWhereThisModelIsRequired = schema.fieldsWhereThisModelIsRequired(mutaction.where.model)
-    val actions                        = fieldsWhereThisModelIsRequired.map(field => mutationBuilder.errorIfNodeIsInRelation(id, field))
-    DBIO.sequence(actions)
-  }
+case class DeleteNodesInterpreter(mutaction: DeleteNodes, shouldDeleteRelayIds: Boolean)(implicit ec: ExecutionContext)
+    extends TopLevelDatabaseMutactionInterpreter {
+
+  def dbioAction(mutationBuilder: JdbcActionsBuilder) =
+    for {
+      ids        <- mutationBuilder.getNodeIdsByFilter(mutaction.model, mutaction.whereFilter)
+      groupedIds = ids.grouped(ParameterLimit.groupSize).toVector
+      _          <- DBIO.seq(groupedIds.map(SharedDelete.checkForRequiredRelationsViolations(mutaction.project, mutaction.model, mutationBuilder, _)): _*)
+      _          <- DBIO.seq(groupedIds.map(mutationBuilder.deleteNodes(mutaction.model, _, shouldDeleteRelayIds)): _*)
+    } yield ManyNodesResult(mutaction, ids.size)
 }
 
 case class NestedDeleteNodeInterpreter(mutaction: NestedDeleteNode, shouldDeleteRelayIds: Boolean)(implicit val ec: ExecutionContext)
@@ -54,7 +60,7 @@ case class NestedDeleteNodeInterpreter(mutaction: NestedDeleteNode, shouldDelete
       childId <- getChildId(mutationBuilder, parentId)
       _       <- mutationBuilder.ensureThatNodesAreConnected(parentField, childId, parentId)
       _       <- performCascadingDelete(mutationBuilder, child, childId)
-      _       <- checkForRequiredRelationsViolations(mutationBuilder, childId)
+      _       <- SharedDelete.checkForRequiredRelationsViolations(mutaction.project, mutaction.relationField.relatedModel_!, mutationBuilder, Vector(childId))
       _       <- mutationBuilder.deleteNodeById(child, childId, shouldDeleteRelayIds)
     } yield UnitDatabaseMutactionResult
   }
@@ -80,10 +86,13 @@ case class NestedDeleteNodeInterpreter(mutaction: NestedDeleteNode, shouldDelete
         }
     }
   }
+}
 
-  private def checkForRequiredRelationsViolations(mutationBuilder: JdbcActionsBuilder, parentId: IdGCValue): DBIO[_] = {
-    val fieldsWhereThisModelIsRequired = mutaction.project.schema.fieldsWhereThisModelIsRequired(mutaction.relationField.relatedModel_!)
-    val actions                        = fieldsWhereThisModelIsRequired.map(field => mutationBuilder.errorIfNodeIsInRelation(parentId, field))
+object SharedDelete {
+  def checkForRequiredRelationsViolations(project: Project, model: Model, mutationBuilder: JdbcActionsBuilder, ids: Vector[IdGCValue])(
+      implicit ec: ExecutionContext): DBIO[_] = {
+    val fieldsWhereThisModelIsRequired = project.schema.fieldsWhereThisModelIsRequired(model)
+    val actions                        = fieldsWhereThisModelIsRequired.map(field => mutationBuilder.errorIfNodesAreInRelation(ids, field))
     DBIO.sequence(actions)
   }
 }
